requirements.txt
# Runtime
flask
webargs
pymongo
PyYaml
jsonpickle

# Develop
click
sphinx
ipython

pytest
pytest-cov
codecov
